Handle overflow warnings for Sigmoid in tests.
In one case it could be fixed by appropriately adjusting the radius.
In all other cases I added a filterwarnings decorator to ignore the
warning as it is not really problematic (the firing rate will be 0
for the Sigmoid neurons because of the overflow which is not that
different from a really small value that we would get with perfect
precision). Because it is not really problematic we could also
selectively disable that warning in the Sigmoid.step_math function,
but I am not sure about the performance impact (two np.seterr calls
each step) and it might still be a useful warning to remind one that
the parameters of the sigmoid neuron might not be well chosen for the
input range.

However, it might be worth considering if it actually makes sense to
always throw Sigmoid neurons into the tests with all the other neuron
types. The Sigmoid neuron type is fundamentally different in that its
intercepts are not actually intercepts, but inflection points.