squash! Handle overflow warnings for Sigmoid in tests.
This removes the individual warnings filters for sigmoid neurons and
automatically installs the filter for all tests with sigmoid neurons.

Unfortunately, setting a mark with pytest.param seems to overwrite all
other marks. Or appends the filterwarnings mark at the end while it
needs to be at the beginning, I'm not sure. At least the existing marks
need to be appended to the new mark when circumventing this issue.